requirements/local.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
pytz==2019.3
sqlparse==0.3.1
